refactor: Add user handler
