+++ internal/handler/auth/auth.go
@@ -1,11 +1,11 @@
 package auth
 
 import (
+	"net/http"
+	"time"
+
 	"github.com/arya237/foodPilot/internal/services"
 	"github.com/arya237/foodPilot/pkg/logger"
-	"net/http"
-	"strconv"
-	"time"
 
 	"github.com/arya237/foodPilot/internal/auth"
 	"github.com/gin-gonic/gin"
@@ -27,8 +27,6 @@ func NewLoginHandler(expiry time.Duration, u services.UserService) *LoginHandler
 
 func RegisterRoutes(group *gin.RouterGroup, loginHandler *LoginHandler) {
 	group.POST("/login", loginHandler.HandleLogin)
-	group.PUT("/autosave", auth.AuthMiddleware(), loginHandler.AutoSave)
-
 }
 
 // ***************** methodes *********************************//
@@ -56,31 +54,4 @@ func (h *LoginHandler) HandleLogin(c *gin.Context) {
 	c.JSON(http.StatusOK, gin.H{"loginResponse": LoginResponse{
 		Token: jwtToken,
 	}, "message": "login successful"})
-}
-
-func (h *LoginHandler) AutoSave(c *gin.Context) {
-
-	req := AutoSaveRequest{}
-	if err := c.ShouldBindJSON(&req); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid request"})
-		return
-	}
-
-	id, exist := c.Get("userID")
-	if !exist {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "user not found"})
-		return
-	}
-	userID, _ := strconv.Atoi(id.(string))
-
-	err := h.UserService.ToggleAutoSave(userID, *req.AutoSave)
-
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "server error!"})
-		return
-	}
-
-	c.JSON(http.StatusOK, AutoSaveResponse{
-		Message: "Auto save updated",
-	})
 }
