+++ internal/handler/user/models.go
@@ -1,0 +1,9 @@
+package user
+
+type AutoSaveRequest struct {
+	AutoSave *bool `json:"autosave" binding:"required"`
+}
+
+type AutoSaveResponse struct {
+	Message string `json:"message"`
+}
