+++ internal/models/rate.go
@@ -3,5 +3,5 @@ package models
 type Rate struct {
 	UserID int
 	FoodID int
-	Score   int
+	Score  int
 }
