+++ internal/services/rate_service.go
@@ -2,10 +2,11 @@ package services
 
 import (
 	"errors"
+	"strconv"
+
 	"github.com/arya237/foodPilot/internal/models"
 	"github.com/arya237/foodPilot/internal/repositories"
 	"github.com/arya237/foodPilot/pkg/logger"
-	"strconv"
 )
 
 type RateFoodService interface {
